Replaced numpy.int with numpy.int32 for compatibility with numpy >= 2.0

+++ krypy/utils.py
@@ -2067,7 +2067,7 @@ class NormalizedRootsPolynomial(object):
         # values
         for j in range(vals.shape[1]):
             sort_tmp = numpy.argsort(numpy.abs(vals[:, j]))
-            sort = numpy.zeros((n,), dtype=numpy.int)
+            sort = numpy.zeros((n,), dtype=numpy.int32)
             mid = int(numpy.ceil(float(n) / 2))
             sort[::2] = sort_tmp[:mid]
             sort[1::2] = sort_tmp[mid:][::-1]
